+++ CHANGES.md
@@ -1,3 +1,8 @@
+0.2.1 (2019-12-06)
+------------------
+- Parallax motion is now stored inside the fitter as a dictionary with `ra_plx` and `dec_plx` keys 
+  so that dec and ra motion cannot be mixed up.
+
 0.2.0 (2019-10-25)
 ------------------
 - Added support for fitting parallaxes, and arbitrarily